quereis

+++ backend/src/db/queries.ts
@@ -13,14 +13,27 @@ export const getUserById = async(id:string) => {
 };
 
 export const updateUser = async(id:string, data: Partial<NewUser>) => {
+    const existingUser = await getUserById(id);
+    if(!existingUser) {
+        throw new Error(`User with  id ${id} not found`);
+    }
     const [user] = await db.update(users).set(data).where(eq(users.id,id)).returning();
     return user;
 };
 
 export const upsertUser = async(data:NewUser) => {
-    const existingUser=await getUserById(data.id)
-    if(existingUser) return updateUser(data.id,data)
-        return createUser(data);
+    // const existingUser=await getUserById(data.id)
+    // if(existingUser) return updateUser(data.id,data)
+    //     return createUser(data);
+    const [user] = await db
+        .insert(users)
+        .values(data)
+        .onConflictDoUpdate({
+            target: users.id,
+            set: data,
+        })
+        .returning();
+    return user;
 };
 
 export const createProduct = async(data: NewProduct) => {
@@ -61,21 +74,33 @@ export const getProductsByUserId = async (userId: string) => {
 };
 
 export const updateProduct = async (id: string, data: Partial<NewProduct>) => {
+    const existingProduct = await getCommentById(id);
+    if (!existingProduct) {
+        throw new Error(`Product with id ${id} not found`);
+    }
     const [product] = await db.update(products).set(data).where(eq(products.id,id)).returning();
     return product;
 };
 
 export const deleteProduct = async (id: string) => {
+    const existingProduct = await getCommentById(id);
+    if (!existingProduct) {
+        throw new Error(`Product with id ${id} not found`);
+    }
     const [product] = await db.delete(products).where(eq(products.id,id)).returning();
     return product;
 };
 
-export const createdComment = async (data: NewComment) => {
+export const createComment = async (data: NewComment) => {
     const [comment] = await db.insert(comments).values(data).returning();
     return comment;
 };
 
 export const deleteComment = async ( id: string ) => {
+    const existingComment = await getCommentById(id);
+    if( !existingComment) {
+        throw new Error(`Comment with id ${id} not found`);
+    }
     const [comment] = await db.delete(comments).where(eq(comments.id,id)).returning();
     return comment;
 };
